\name{moPMR-package}
\alias{moPMR-package}
\alias{moPMR}
\docType{package}
\title{\packageTitle{moPMR}}
\description{moPMR is an R package for efficient statistical inference of two-sample MR analysis for multiple outcome traits. It can account for the correlated instruments and the horizontal pleiotropy, and can provide the accurate estimates of both causal effect and horizontal pleiotropy effect as well as the two corresponding p values.There are two main functions in moPMR package, one is moPMR_individual for individual level data, the other is moPMR_summary for summary data.}
\details{
 \tabular{ll}{
Package: \tab moPMR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-08-07\cr
License: \tab GPL-3\cr
}
}
\author{
   Zhongshang Yuan, Xiang Zhou
  
  Maintainer: Zhongshang Yuan <yuanzhongshang@sdu.edu.cn>
}
\references{
Yuan, Z., Zhu, H., Zeng, P., Yang, S., Sun, S., Yang, C., Liu, J., and Zhou, X. (2019). Testing and controlling for horizontal pleiotropy with the probabilistic mendelian randomization in transcriptome-wide association studies. bioRxiv pp. 691014.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}